Feat: Complete Decoder implementation, refactor Encoder, add TypeConverter functionality
- Refactored `StructEncoder` for better maintainability and performance.
- Fully implemented `StructDecoder` for decoding structures from byte arrays.
- Introduced `TypeConverterHelper` functionality for robust type conversion between various types, including null value handling, primitives, and complex objects.
- Ensured consistent handling of null and default values across both Encoder and Decoder components.
- Added comprehensive tests for all new features, including type conversion validation and edge cases.

+++ src/main/java/net/deanly/structlayout/type/basic/UInt8Layout.java
@@ -2,7 +2,7 @@ package net.deanly.structlayout.type.basic;
 
 import net.deanly.structlayout.Layout;
 
-public class UInt8Layout extends Layout<Integer> {
+public class UInt8Layout extends Layout<Short> {
 
     public UInt8Layout(String property) {
         super(1, property); // 1 byte
@@ -13,15 +13,15 @@ public class UInt8Layout extends Layout<Integer> {
     }
 
     @Override
-    public Integer decode(byte[] data, int offset) {
+    public Short decode(byte[] data, int offset) {
         if (data == null) {
             throw new IllegalArgumentException("Data cannot be null.");
         }
-        return data[offset] & 0xFF; // 부호 없는 8비트 정수
+        return (short) (data[offset] & 0xFF); // 부호 없는 8비트 정수
     }
 
     @Override
-    public byte[] encode(Integer value) {
+    public byte[] encode(Short value) {
         if (value == null || value < 0 || value > 255) {
             throw new IllegalArgumentException("Value must be in the range 0 to 255.");
         }
